+++ src/molecules/SkeletonChart/index.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import styles from './SkeletonChart.module.css';
 
-import Skeleton from '../../atoms/Skeleton/Skeleton';
+import Skeleton from '../../atoms/Skeleton';
 
 export default function SkeletonChart({ columns = 16 }) {
 	const [randomColumnHights] = useState(Array.from({ length: columns }, () => Math.floor(Math.random() * 240) + 40));