+++ src/organisms/EventList/EventList.js
@@ -4,9 +4,9 @@ import React, { useEffect, useState } from 'react';
 import { useQuery, gql } from '@apollo/client';
 
 import EventCard from '../../molecules/EventCard/EventCard';
-import Message from '../../atoms/Message/Message';
+import Message from '../../atoms/Message';
 
-import Skeleton from '../../atoms/Skeleton/Skeleton';
+import Skeleton from '../../atoms/Skeleton';
 
 const EVENTS = gql`
 	query events (