+++ src/pages/SmartMeterSettings/SmartMeterSettings.js
@@ -2,8 +2,8 @@ import React from 'react';
 
 import Settings from '../../templates/Settings/Settings';
 
-import Skeleton from '../../atoms/Skeleton/Skeleton';
-import Message from '../../atoms/Message/Message';
+import Skeleton from '../../atoms/Skeleton';
+import Message from '../../atoms/Message';
 
 import { DateTime } from 'luxon';