+++ src/templates/Settings/Settings.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './Settings.module.css';
 
-import Header from '../../organisms/Header/Header';
+import Header from '../../organisms/Header';
 import SettingsHeader from '../../molecules/SettingsHeader/SettingsHeader';
 
 export default function Default({ children, title = '', backButton = true }) {