+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.14.0] - 2021-11-08
+### Added
+- Electricity year widget.
+- User header modal.
+
 ## [1.13.0] - 2021-05-24
 ### Changed
 - **Needs API version >= 2.0** Refactor for breaking changes on the API.